+++ app/javascript/dashboard/store/modules/integrations.js
@@ -2,7 +2,13 @@
 import * as MutationHelpers from 'shared/helpers/vuex/mutationHelpers';
 import * as types from '../mutation-types';
 import IntegrationsAPI from '../../api/integrations';
-import { throwErrorMessage } from 'dashboard/store/utils/api';
+import {
+  parseAPIErrorResponse,
+  throwErrorMessage,
+} from 'dashboard/store/utils/api';
+
+const normalizeApiError = error =>
+  Object.assign(new Error(parseAPIErrorResponse(error)), error);
 
 const state = {
   records: [],
@@ -112,7 +118,7 @@ export const actions = {
       commit(types.default.SET_INTEGRATIONS_UI_FLAG, { isFetchingItem: false });
     } catch (error) {
       commit(types.default.SET_INTEGRATIONS_UI_FLAG, { isFetchingItem: false });
-      throw new Error(error);
+      throw normalizeApiError(error);
     }
   },
   createHook: async ({ commit }, hookData) => {
@@ -123,7 +129,7 @@ export const actions = {
       commit(types.default.SET_INTEGRATIONS_UI_FLAG, { isCreatingHook: false });
     } catch (error) {
       commit(types.default.SET_INTEGRATIONS_UI_FLAG, { isCreatingHook: false });
-      throw new Error(error);
+      throw normalizeApiError(error);
     }
   },
   updateHook: async ({ commit }, { hookId, hookData }) => {
@@ -134,7 +140,7 @@ export const actions = {
       commit(types.default.SET_INTEGRATIONS_UI_FLAG, { isUpdatingHook: false });
     } catch (error) {
       commit(types.default.SET_INTEGRATIONS_UI_FLAG, { isUpdatingHook: false });
-      throw new Error(error);
+      throw normalizeApiError(error);
     }
   },
   deleteHook: async ({ commit }, { appId, hookId }) => {
@@ -145,7 +151,7 @@ export const actions = {
       commit(types.default.SET_INTEGRATIONS_UI_FLAG, { isDeletingHook: false });
     } catch (error) {
       commit(types.default.SET_INTEGRATIONS_UI_FLAG, { isDeletingHook: false });
-      throw new Error(error);
+      throw normalizeApiError(error);
     }
   },
   runHookSync: async ({ commit }, hookId) => {